+++ changelog.md
@@ -5,6 +5,11 @@ For information on Anvato Player changes, see their [changelog](https://dev.anva
 
 ## Version 3
 
+#### [3.1.0]
+- Allow for individual player initial plugin configs
+- Modify and return the initial config in the beforeVideoLoad binding
+- Updated the common comScore plugin to use the latest script
+
 #### [3.0.1]
 - Ensure videoPlayerLoad action fires only once per session per video player